pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contsg"
version = "0.1.0"
description = "Conditional Time Series Generation Benchmark"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ConTSG Team"}
]
keywords = ["time-series", "generation", "benchmark", "deep-learning", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "gradio>=4.0.0",
    "plotly>=5.18.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "tensorboard>=2.14.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.0.0",
]
full = [
    "transformers>=4.30.0",
    "einops>=0.6.0",
    "linear-attention-transformer>=0.19.0",
    "opt_einsum>=3.3.0",
    "timm>=0.9.0",
    "datasets>=2.14.0",
    "sentence-transformers>=2.2.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "tslearn>=0.6.0",
    "pykeops>=2.1.0",
    "omegaconf>=2.3.0",
]

[project.scripts]
contsg = "contsg.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/contsg"
Documentation = "https://github.com/your-org/contsg#readme"
Repository = "https://github.com/your-org/contsg"

[tool.setuptools.packages.find]
where = ["."]
include = ["contsg*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=contsg"
